+++ Toggly.FeatureManagement.Node/toggly-fastify/src/plugin.ts
@@ -136,6 +136,9 @@ const togglyPluginAsync: FastifyPluginAsync<TogglyFastifyConfig> = async (
       isFeatureOff: (featureKey: string) => fastifyClient!.isFeatureOff(featureKey, context),
       evaluateFeatureGate: (featureKeys, requirement, negate) =>
         fastifyClient!.evaluateFeatureGate(featureKeys, requirement, negate, context),
+      getVariant: (featureKey: string) => fastifyClient!.getVariant(featureKey, context),
+      getVariantValue: (featureKey: string) =>
+        fastifyClient!.getVariantValue(featureKey, context),
     }
   })