+++ Toggly.FeatureManagement.Node/toggly-koa/src/middleware.ts
@@ -130,6 +130,9 @@ export function togglyMiddleware(config: TogglyKoaConfig): Middleware {
       isFeatureOff: (featureKey: string) => koaClient!.isFeatureOff(featureKey, evalContext),
       evaluateFeatureGate: (featureKeys, requirement, negate) =>
         koaClient!.evaluateFeatureGate(featureKeys, requirement, negate, evalContext),
+      getVariant: (featureKey: string) => koaClient!.getVariant(featureKey, evalContext),
+      getVariantValue: (featureKey: string) =>
+        koaClient!.getVariantValue(featureKey, evalContext),
     }
 
     await next()